requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.8
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
